Bindicator v2 based on hacs_waste_collection_schedule and config.
Use hacs_waste_collection_schedule as the data source for the bin
collections schedules.

An Ansible role is created to install hacs_waste_collection_schedule
directly from GitHub instead of using HACS n.b. once cloned need to
copy custom_component directory directly to target as
Home Assistant appears to not follow symlinks.

Sensors are created for each of the blue and black bins with the
values of how many days until their next collection - 1 means
tomorrow.

Define a bindicator automation based entirely on built-ins -
trigger at 4pm (changed from 5pm) and then if < 2 days to black bin
light living room lamp up magenta, else if < 2 days to blue bin
light up blue.

Job done.

--

Replaced the POST calls to Home Assistant HTTP API for

- Configuration reload
- Restart (core)

as the restart would kill the server before the request could
complete (lol) and the as far as I can tell the reload wasn't
effective for what I was trying to use it for (reload automations
etc.).

Replaced with "ha core restart" command to do a restart by SSH,
which is a bit heavy-handed for reloading configuration but it
doesn't take that long really and it gets the job done.

--

Fiddled with the Ansible layout a bit - inventory file made into a
directory with auto-loaded files so that we can use

ansible-playbook -i inventory/local

without needing to specify .yaml at the end which in my opinion
is nicer and without having to truncate the file extension and mess
about with specific file associations in the editor.